llm_service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
mistralai
requests